# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Margaz Tank Telemetry API
  version: 1.0.0
  description: "API documentation for the Margaz Tank Telemetry backend. Frontend: https://margaz.netlify.app"
  license:
    name: Proprietary
    url: https://margaz.netlify.app
servers:
  - url: http://63.181.47.189
    description: Lightsail direct
...
security:
  - bearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: admin@example.com
              password: "secret123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              example:
                message: "Giriş başarılı"
                token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                user:
                  id: "b7b22a4a-3e6e-4df2-9fa0-1b9d8c9a0a11"
                  email: admin@example.com
                  name: "Admin"
                  role: ADMIN
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: viewer@example.com
              password: "secret123"
              name: "Viewer"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
            example:
              currentPassword: "oldpass123"
              newPassword: "newpass123"
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required: [email, password]
      example:
        email: admin@example.com
        password: "secret123"
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required: [email, password]
      example:
        email: viewer@example.com
        password: "secret123"
        name: "Viewer"
    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
      required: [currentPassword, newPassword]
      example:
        currentPassword: "oldpass123"
        newPassword: "newpass123"
    AuthUser:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [ADMIN, VIEWER]
      required: [id, email, role]
    AuthResponse:
      type: object
      properties:
        message:
          type: string
        token:
          type: string
        user:
          $ref: '#/components/schemas/AuthUser'
      required: [message, token, user]
      example:
        message: "Giriş başarılı"
        token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          id: "b7b22a4a-3e6e-4df2-9fa0-1b9d8c9a0a11"
          email: admin@example.com
          name: "Admin"
          role: ADMIN
    MeResponse:
      type: object
      properties:
        user:
          allOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: object
              properties:
                createdAt:
                  type: string
                  format: date-time
      required: [user]
...